Improve insert teacher replacing person entity
