+++ src/main/java/com/react/pnld/mappers/TeacherMapper.java
@@ -1,14 +1,14 @@
 package com.react.pnld.mappers;
 
-import com.react.pnld.model.Person;
+import com.react.pnld.model.Teacher;
 import org.apache.ibatis.annotations.Insert;
 import org.apache.ibatis.annotations.Result;
 import org.apache.ibatis.annotations.Results;
 import org.apache.ibatis.annotations.Select;
 
-public interface PersonMapper {
+public interface TeacherMapper {
 
-    @Select("SELECT * FROM pnld.persona where rut=#{rut} AND correo=#{email}")
+    @Select("SELECT * FROM pnld.docente where rut=#{rut} AND correo=#{email}")
     @Results({
             @Result(property = "id", column = "id"),
             @Result(property = "firstName", column = "nombre"),
@@ -17,13 +17,13 @@ public interface PersonMapper {
             @Result(property = "rut", column = "rut"),
             @Result(property = "email", column = "correo"),
             @Result(property = "genderId", column = "id_genero")})
-    Person getPerson(String rut, String email);
+    Teacher getTeacher(String rut, String email);
 
-    @Insert("INSERT INTO pnld.persona (nombre, apellido_paterno, apellido_materno, rut," +
-            "correo, id_genero) VALUES (#{firstName},#{lastName},#{motherLastName},#{rut}," +
+    @Insert("INSERT INTO pnld.docente (id, nombre, apellido_paterno, apellido_materno, rut," +
+            "correo, id_genero) VALUES (#{id}, #{firstName},#{lastName},#{motherLastName},#{rut}," +
             "#{email},#{genderId});")
-    int insertPerson(Person person);
+    int insertTeacher(Teacher teacher);
 
-    @Insert("INSERT INTO pnld.docente SELECT id FROM pnld.persona WHERE rut=#{personRut};")
-    int insertTeacherByPersonRut(String personRut);
+    @Select("SELECT nextval(pg_get_serial_sequence('pnld.docente', 'id'));")
+    int getNextTeacherId();
 }
